refactor(bridge): isolate secure download storage
